Cleanup and provide the original enhance procedure

+++ audio.py
@@ -6,16 +6,17 @@ from numpy.typing import NDArray
 
 
 def read(path: str, sr: int) -> NDArray:
-	
+
     samples, samplerate = soundfile.read(path, always_2d=True)
 
     if samplerate != sr:
         samples = resampy.resample(samples, samplerate, sr)
         samplerate = sr
 
-    samples = samples.astype(np.float32).T
-	
+    samples = samples.T
+
     return samples, samplerate
+
 
 def write(path: str, sr: int, samples: NDArray):
